+++ React/Base/RCTModuleMethod.m
@@ -480,12 +480,12 @@ SEL RCTParseMethodSignature(NSString *methodSignature, NSArray<RCTMethodArgument
         expectedCount -= 2;
       }
 
-      RCTLogError(@"%@.%@ was called with %zd arguments but expects %zd arguments. "
+      RCTLogError(@"%@.%@ was called with %lld arguments but expects %lld arguments. "
                   @"If you haven\'t changed this method yourself, this usually means that "
                   @"your versions of the native code and JavaScript code are out of sync. "
                   @"Updating both should make this error go away.",
                   RCTBridgeModuleNameForClass(_moduleClass), _JSMethodName,
-                  actualCount, expectedCount);
+                  (long long)actualCount, (long long)expectedCount);
       return nil;
     }
   }
